chore: rename websocket tutorial folder
