+++ .github/workflows/tests.yaml
@@ -35,7 +35,7 @@ jobs:
           - "towncrier"
           - "look"
           - "asgilook"
-          - "wslook"
+          - "ws_tutorial"
           - "check_vendored"
           - "twine_check"
           - "daphne"
