+++ examples/ws_tutorial/ws_tutorial/static/index.html
