--- examples/wslook/.coveragerc
@@ -1,4 +1,0 @@
-[run]
-omit =
-    examples/wslook/wslook/client.py
-    examples/wslook/wslook/reports_client.py
