+++ tox.ini
@@ -442,21 +442,21 @@ commands =
         {toxinidir}/examples/asgilook/tests/
 
 # --------------------------------------------------------------------
-# WebSockets tutorial ("wslook") tests
+# WebSockets tutorial ("ws_tutorial") tests
 # --------------------------------------------------------------------
 
-[testenv:wslook]
+[testenv:ws_tutorial]
 basepython = python3.12
 deps =
-    -r{toxinidir}/examples/wslook/requirements/app
-    -r{toxinidir}/examples/wslook/requirements/test
+    -r{toxinidir}/examples/ws_tutorial/requirements/app
+    -r{toxinidir}/examples/ws_tutorial/requirements/test
 commands =
     pytest \
-        --cov wslook \
-        --cov-config {toxinidir}/examples/wslook/.coveragerc \
+        --cov ws_tutorial \
+        --cov-config {toxinidir}/examples/ws_tutorial/.coveragerc \
         --cov-fail-under 100 \
         --cov-report term-missing \
-        {toxinidir}/examples/wslook/tests/
+        {toxinidir}/examples/ws_tutorial/tests/
 
 # --------------------------------------------------------------------
 # Ecosystem
